refactor contactDetail
